More control of fzf window position at the bottom of frame
Support selecting the bottom-most window and/or a bottom root-window.

The original idea came from m-milek for using a window at the bottom.
The current code is an evolution of the idea from m-milek, although I think
more cleanup is required to simplify the user-options and make the control of
the fzf window a little bit more flexible: use top or bottom root window when
the splitting would create a window too small.

See https://github.com/bling/fzf.el/pull/122

The code is not yet integrated into https://github.com/bling/fzf.el

So, for now I'm putting it inside my fork.